requirements.txt
discord.py
python-dotenv
sqlalchemy
asyncpg
pydantic-settings
